// projects/ui-smartcontact/src/lib/core/types/theme-component.types.ts
export type ScSeverity = 'primary' | 'secondary' | 'success' | 'info' | 'warn' | 'warning' | 'danger' | 'contrast';

export type ScComponentSize = 'sm' | 'md' | 'lg';

export type ScInputVariant = 'outlined' | 'filled';

export type ScOverlayPosition = 'left' | 'right' | 'top' | 'bottom';

export type ScDialogPosition =
    | 'center'
    | 'top'
    | 'bottom'
    | 'left'
    | 'right'
    | 'topleft'
    | 'topright'
    | 'bottomleft'
    | 'bottomright';

export type ScAvatarSize = 'normal' | 'large' | 'xlarge';

export type ScAvatarShape = 'square' | 'circle';

export type ScSkeletonShape = 'rectangle' | 'circle';

export type ScSkeletonAnimation = 'wave' | 'none';

export type ScProgressBarMode = 'determinate' | 'indeterminate';